完善信息注册和状态控制

+++ src/main/java/com/by/dao/WxUserLoginMapper.java
@@ -6,5 +6,5 @@ import com.by.bean.WxUser;
 public interface WxUserLoginMapper {
     WxUser selectByOpenid(String openid);
 
-    int insertAndReturnPK(User record);
+    long insertAndReturnPK(User record);
 }
